refactor and fix: move InputSchema tests to serialization
